refactor(metrics): Rename metric to metrics
